# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Chess Game API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type GameStatus = (typeof GameStatus)[keyof typeof GameStatus];

export const GameStatus = {
  active: "active",
  white_wins: "white_wins",
  black_wins: "black_wins",
  draw: "draw",
  stalemate: "stalemate",
} as const;

export type GameCurrentTurn =
  (typeof GameCurrentTurn)[keyof typeof GameCurrentTurn];

export const GameCurrentTurn = {
  white: "white",
  black: "black",
} as const;

export interface Game {
  id: string;
  whitePlayerName: string;
  blackPlayerName: string;
  /** FEN notation of current board state */
  fen: string;
  status: GameStatus;
  currentTurn: GameCurrentTurn;
  moveCount: number;
  createdAt: string;
  updatedAt: string;
}

export interface Move {
  id: string;
  gameId: string;
  moveNumber: number;
  from: string;
  to: string;
  piece: string;
  captured?: string | null;
  promotion?: string | null;
  /** Standard Algebraic Notation */
  san: string;
  /** FEN after this move */
  fen: string;
  createdAt: string;
}

export interface CreateGameRequest {
  whitePlayerName: string;
  blackPlayerName: string;
}

export interface MakeMoveRequest {
  /** Source square (e.g. e2) */
  from: string;
  /** Target square (e.g. e4) */
  to: string;
  /** Promotion piece (q, r, b, n) */
  promotion?: string | null;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}