# lib/api-zod/src/generated/types/game.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Chess Game API
 * OpenAPI spec version: 0.1.0
 */
import type { GameCurrentTurn } from "./gameCurrentTurn";
import type { GameStatus } from "./gameStatus";

export interface Game {
  id: string;
  whitePlayerName: string;
  blackPlayerName: string;
  /** FEN notation of current board state */
  fen: string;
  status: GameStatus;
  currentTurn: GameCurrentTurn;
  moveCount: number;
  createdAt: string;
  updatedAt: string;
}