# lib/api-zod/src/generated/types/makeMoveRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Chess Game API
 * OpenAPI spec version: 0.1.0
 */

export interface MakeMoveRequest {
  /** Source square (e.g. e2) */
  from: string;
  /** Target square (e.g. e4) */
  to: string;
  /** Promotion piece (q, r, b, n) */
  promotion?: string | null;
}