# lib/api-zod/src/generated/types/move.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Chess Game API
 * OpenAPI spec version: 0.1.0
 */

export interface Move {
  id: string;
  gameId: string;
  moveNumber: number;
  from: string;
  to: string;
  piece: string;
  captured?: string | null;
  promotion?: string | null;
  /** Standard Algebraic Notation */
  san: string;
  /** FEN after this move */
  fen: string;
  createdAt: string;
}